LocalFunction_Project/requirements.txt
azure-functions
numpy==1.18.5
pandas==1.0.4
matplotlib==3.2.1